[gas] Update cost tables
Updates cost tables to be more resilient to adverserial workloads.

The current tiers have remained unchanged, but the tail-end of the
tables is increased to guard against large instruction/large allocation.

This also adds versioning for the cost table, and adds support for
running transactions `unmetered` in the transactional test runner.